Move the Chamfer Distance to a new file & Check the shape of the input point cloud.
